Removes two tests that were incorrectly passing.
`expectAsync`  wasn't being awaited, so these tests passed even though they should have both failed.
However, the tests themselves were not testing something reasonable.

In the case of `throws on wrong message`, it was trying to test that that it would throw if the wrong message was supplied, but recover will essentially always succeed and just spit out the wrong address.

In the case of `throws on wrong message` I think it was trying to test that the particular r/recoveryParam wasn't valid but it turns out most pairs are valid so this is another thing that is very hard to test.
In the process of researching this, I found that I wasn't checking that `r` and `s` were in range when recovering (unlike verifying), so I fixed that and updated the test to validate.

Unfortunately, ttypescript doesn't appear to work in NodeJS 18 anymore so I have also introduced a dockerized dev container setup as fixing the ttypescript issue is a bigger change.